update key map for neo-tree from nvim-tree, initial plugin migration for lazy.nvim, reconfigure vimwiki
