+++ lua/mappings.lua
@@ -3,7 +3,7 @@ local map = vim.api.nvim_set_keymap
 -- general
 map('', '<C-f>', ':LspZeroFormat<CR>', {})
 map('n', '<leader>f', '<Plug>(easymotion-bd-f)', {})
-map('n', '<leader>o', ':NvimTreeToggle<CR>', {})
+map('n', '<leader>o', ':Neotree toggle reveal float<CR>', {})
 map('n', '<leader>p', ':Telescope find_files<cr>', {})
 map('n', '<leader>gf', ':GF?<CR>', {})
 map('n', '<leader>/', ':Telescope live_grep<cr>', {})
