+++ lua/plugins/barbar.lua
@@ -1,0 +1,8 @@
+return { -- buffer tabs
+    'romgrk/barbar.nvim',
+    dependencies = {
+        'lewis6991/gitsigns.nvim', -- OPTIONAL: for git status
+        'nvim-tree/nvim-web-devicons', -- OPTIONAL: for file icons
+    },
+    -- config = get_config('barbar'),
+}
