+++ lua/plugins/comment.lua
@@ -1,0 +1,4 @@
+return { -- comment motion
+    'numToStr/Comment.nvim',
+    -- config = get_config('Comment'),
+}
