+++ lua/plugins/neo-tree.lua
@@ -1,0 +1,22 @@
+return {
+    "nvim-neo-tree/neo-tree.nvim",
+    branch = "v3.x",
+    dependencies = {
+      "nvim-lua/plenary.nvim",
+      "nvim-tree/nvim-web-devicons", -- not strictly required, but recommended
+      "MunifTanjim/nui.nvim",
+      "3rd/image.nvim", -- Optional image support in preview window: See `# Preview Mode` for more information
+    },
+    config = {
+        close_if_last_window = true,
+        filesystem = {
+            filtered_items = {
+                hide_dotfiles = false,
+                hide_gitignored = false,
+            },
+            follow_current_file = {
+                enabled = true,
+            },
+        },
+    },
+}
